+++ catalog/model/extension/openbay/amazonus_product.php
@@ -97,7 +97,7 @@ class ModelExtensionOpenBayAmazonusProduct extends Model {
 		$sql_values = array();
 
 		foreach ($data as $product) {
-			$sql_values[] = " ('" . $this->db->escape($product['sku']) . "', " . (int)$product['quantity'] . ", '" . $this->db->escape($product['asin']) . "', " . (double)$product['price'] . ") ";
+			$sql_values[] = " ('" . $this->db->escape($product['sku']) . "', " . (int)$product['quantity'] . ", '" . $this->db->escape($product['asin']) . "', " . (float)$product['price'] . ") ";
 		}
 
 		$sql .= implode(',', $sql_values);
